% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APICalls.R
\name{apiDataSets}
\alias{apiDataSets}
\title{Get DataSets}
\usage{
apiDataSets(usr = "Demo", key = "Demo")
}
\arguments{
\item{usr}{user name}

\item{key}{API key}
}
\description{
Returns information about the DataSets available in the SoilDataFederator.
}
\examples{
apiDataSets()

}
