% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APICalls.R
\name{apiGetSoilData}
\alias{apiGetSoilData}
\title{Get SoilData}
\usage{
apiGetSoilData(observedProperty = NULL, PropertyGroup = NULL,
  DataSet = NULL, bbox = NULL, numToReturn = NULL, usr = "Demo",
  key = "Demo")
}
\arguments{
\item{observedProperty}{(Optional) Specify the soil data property/s to return. It should be a single observedProperty code or a semi-colon delimited text string of observedProperty codes.}

\item{PropertyGroup}{(Optional)  Extract data for a defined group of soil properties.}

\item{DataSet}{(Optional) Filter the data returned to a specific set of DataSets. It should be a single DataSet code or a semi-colon delimited text string of DataSet codes. Default = All DataSets}

\item{bbox}{(Optional) The rectangular bounding box of the area in the form minx;maxx;miny;maxy - semicolon delimited}

\item{numToReturn}{(Optional) The number of records to be returned. Default = All}

\item{usr}{(Required) User name for accessing the API. To register for an API key go to - https://shiny.esoil.io/SoilDataFederator/Register/ You can use usr=Demo & key=Demo but only the first 5 records will be returned}

\item{key}{(Required) API key for accessing the API.}
}
\description{
Returns soil observation data
}
\examples{
apiGetSoilData()

}
