% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APICalls.R
\name{apiProperties}
\alias{apiProperties}
\title{Get Properties}
\usage{
apiProperties(PropertyGroup = NULL, verbose = TRUE)
}
\arguments{
\item{PropertyGroup}{(Optional) return just the properties for a given PropertyGroup. Default = All}

\item{verbose}{(Optional) return just the property codes or the full descriptions. Default = True}
}
\description{
Returns a listing of the available properties
}
\examples{
apiProperties(PropertyGroup=NULL, verbose=TRUE)

}
