% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/APICalls.R
\name{apiPropertyGroups}
\alias{apiPropertyGroups}
\title{Get PropertyGroups}
\usage{
apiPropertyGroups(PropertyGroup = NULL, verbose = TRUE)
}
\description{
Returns a listing of the available Property Groups
}
\examples{
apiPropertyGroups()

}
