% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMassage.R
\name{makeLocations}
\alias{makeLocations}
\title{Return locations of samples}
\usage{
makeLocations(soilDF, drawit = F)
}
\arguments{
\item{soilDF}{Dataframe to generate site locations for}
}
\description{
Takes the standard SoilDataFederator dataframe returned by getSoilData function and generates a spatialPointsDataFrame .
}
\examples{
makeLocations(soilDF=DF)
}
\keyword{spatialPointsDataFrame}
