% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataMassage.R
\name{mergeObservedProperties}
\alias{mergeObservedProperties}
\title{Merge ObservedProperties}
\usage{
mergeObservedProperties(soilDF)
}
\arguments{
\item{soilDF}{Dataframe to transpose}
}
\description{
Takes the standard SoilDataFederator dataframe returned by getSoilData function and transposes the multiple observed propeties into individual columns.
}
\examples{
mergeObservedProperties(obsPropDF=DF)
}
\keyword{Transform}
