% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeafletMaps.R
\name{showLeafletMap}
\alias{showLeafletMap}
\title{Draw a interactive leaflet map of soil site locations}
\usage{
showLeafletMap(soilDF)
}
\arguments{
\item{soilDF}{Dataframe to map}
}
\description{
Takes the standard SoilDataFederator dataframe returned by getSoilData function and shows the locations in a leaflet map.
}
\examples{
showLeafletMap(soilDF=DF)
}
\keyword{Map}
