% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummariseData.R
\name{summariseContinuousSoilData}
\alias{summariseContinuousSoilData}
\title{Summarises values in a SoilData dataframe}
\usage{
summariseContinuousSoilData(soilDF, outputs = "All")
}
\arguments{
\item{soilDF}{Dataframe to generate site locations for}
}
\description{
Returns a list of various summary statistics and charts.
}
\examples{
summariseContinuousSoilData(soilDF=DF)
}
